fix: 1-list_len.c

+++ 0x12-singly_linked_lists/1-list_len.c
@@ -2,14 +2,14 @@
 #include "lists.h"
 
 /**
- * list_len = return the number of elements in a linked list
+ * list_len - return the number of elements in a linked list
  * @h: the head of list
  *
  * Return: The number of element
  */
 size_t list_len(const list_t *h)
 {
-	int n = 0;
+	size_t n = 0;
 
 	while (h)
 	{
